refactor (LeagueUtil) split off SummonerUtil
