+++ lib/EndpointUtil.js
@@ -1,5 +1,5 @@
 const ParameterError = require('./errors/ParameterError');
-const LeagueUtil = require('./LeagueUtil');
+const SummonerUtil = require('./util/SummonerUtil');
 const RateLimiter = require('./RateLimiter');
 const RegionAndPlatformUtil = require('./util/RegionAndPlatformUtil');
 
@@ -46,10 +46,10 @@ class EndpointUtil {
 		}
 	}
 	static throwIfNameIsInvalid(name) {
-		if (!LeagueUtil.validateSummonerNameInputLength(name)) {
+		if (!SummonerUtil.validateSummonerNameInputLength(name)) {
 			throw new Error(`Summoner name "${name}" is too long (16 characters max, found ${name.length})`);
 		}
-		const invalidCharactersMatch = LeagueUtil.validateSummonerNameInputCharacters(name);
+		const invalidCharactersMatch = SummonerUtil.validateSummonerNameInputCharacters(name);
 		if (invalidCharactersMatch) {
 			throw new Error(`Summoner name "${name}" contains invalid characters: ${invalidCharactersMatch.join(' | ')}`);
 		}
