+++ lib/endpoints/SummonerEndpoint.js
@@ -1,6 +1,6 @@
 const Endpoint = require ('../Endpoint');
 const EndpointUtil = require('../EndpointUtil');
-const LeagueUtil = require('../LeagueUtil');
+const SummonerUtil = require('../util/SummonerUtil');
 
 /**
  * Endpoint to receive Summoner information
@@ -23,7 +23,7 @@ class SummonerEndpoint extends Endpoint {
 		EndpointUtil.throwIfParamIsArray(name, 'name');
 		EndpointUtil.throwIfNameIsInvalid(name, 'name');
 
-		let normalizedName = LeagueUtil.normalizeSummonerName(name);
+		let normalizedName = SummonerUtil.normalizeSummonerName(name);
         return this.executingRequest(`/by-name/${normalizedName}`,  platformIdOrRegion);
     }
